Check table as independent script (#8976)
